requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
PyQt6-WebEngine==6.6.0
PyQt6-WebEngine-Qt6==6.6.0
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.3.2
folium==0.15.1
geopy==2.4.1
shapely==2.0.2
pyproj==3.6.1
cryptography==41.0.7
python-dateutil==2.8.2
pytz==2023.3.post1
tqdm==4.66.1
openpyxl==3.1.2
xlrd==2.0.1
xlwt==1.3.0
pyinstaller==6.3.0